Revert "Extract client to its own class"
This refactor was not required for the functionality, undoing it.

This reverts commit 094b03fcfdf3a63fa06b286c32b5fb5f5d4274c8.
